--- source/watch/index.ts
@@ -1,1 +1,0 @@
-export { Watcher, type WatchEventHandler as WatcherCallback, type WatcherOptions } from "./Watcher.js";
