+++ source/watcher/index.ts
@@ -1,0 +1,1 @@
+export { Watcher, type WatchEventHandler, type WatcherOptions } from "./Watcher.js";
